router/requirements.txt
# Multi-Protocol IoT Network Router Dependencies
# Python packages required for the Raspberry Pi router

# Cryptography for secure handshake and encryption
pycryptodome>=3.15.0
cryptography>=3.4.8

# BLE support for Raspberry Pi
bleak>=0.19.0

# System and network utilities
psutil>=5.8.0
psutil>=5.8.0
netifaces>=0.11.0

# Web interface dependencies
flask>=2.0.0
flask-socketio>=5.0.0
eventlet>=0.33.0

# Database support for persistence
sqlite3

# Configuration management
pyyaml>=6.0
configparser

# Logging enhancements
colorlog>=6.6.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.18.0